- Changed: SpellIDs are now limited to ushort.Max. - Bugfix: Already compressed files will no longer be recompressed with mailmgr if they exist in the folder. - Updated: The DB Converter to handle the Spell ID conversion, among other improvements.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/EtaewMajor@1143 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/spells/Spell.cs
@@ -305,7 +305,7 @@ namespace DOL.GS
 		}
 
 		public Spell(DBSpell dbspell, int requiredLevel)
-			: base(dbspell.Name, (ushort)dbspell.SpellID, requiredLevel)
+			: base(dbspell.Name, dbspell.SpellID, requiredLevel)
 		{
 
 			m_description = dbspell.Description;
